Add locationHash
